grpcio_Game/requirements.txt
grpcio==1.46.3
grpcio-tools==1.46.3
redis==4.3.1